requirements.txt
# 核心机器学习库
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0

# 生物信息学库
biopython>=1.81

# 自然语言处理
transformers>=4.32.0

# 数据处理和可视化
matplotlib>=3.7.2
seaborn>=0.12.2
tqdm>=4.65.0

# 模型序列化
joblib>=1.2.0

# 其他工具
argparse  # 内置库
json      # 内置库
datetime  # 内置库
os        # 内置库
sys       # 内置库
